app/requirements.txt
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
python-dotenv==1.0.1
